[BE-125] refactor: 구간 삭제 에러케이스 세분화

+++ Ticket-Api/src/main/java/com/jnu/ticketapi/api/sector/service/SectorDeleteUseCase.java
@@ -3,6 +3,7 @@ package com.jnu.ticketapi.api.sector.service;
 
 import com.jnu.ticketcommon.annotation.UseCase;
 import com.jnu.ticketdomain.domains.events.domain.Sector;
+import com.jnu.ticketdomain.domains.events.exception.CannotUpdatePublishEventException;
 import com.jnu.ticketdomain.domains.events.out.SectorLoadPort;
 import com.jnu.ticketdomain.domains.events.out.SectorRecordPort;
 import com.jnu.ticketdomain.domains.registration.out.RegistrationRecordPort;
@@ -21,7 +22,9 @@ public class SectorDeleteUseCase {
     @Transactional
     public void execute(Long sectorId) {
         // to Sector List
-        Sector sector = sectorLoadPort.findByIdAndPublishIsFalse(sectorId);
+        Sector sector = sectorLoadPort.findById(sectorId);
+        if (Boolean.TRUE.equals(sector.getEvent().getPublish()))
+            throw CannotUpdatePublishEventException.EXCEPTION;
         registrationRecordPort.deleteBySector(sector);
         sectorRecordPort.delete(sector);
     }
